# modules/apps/object/object-admin-rest-impl/rest-openapi.yaml
components:
    schemas:
...
        ObjectRelationship:
            properties:
                actions:
                    additionalProperties:
                        additionalProperties:
                            type: string
                        type: object
                    readOnly: true
                    type: object
                id:
                    format: int64
                    readOnly: true
                    type: integer
                label:
                    additionalProperties:
                        type: string
                    type: object
                name:
                    type: string
                objectDefinitionId1:
                    format: int64
                    type: integer
                objectDefinitionId2:
                    format: int64
                    type: integer
                objectDefinitionName1:
                    type: string
                objectDefinitionName2:
                    type: string
                type:
                    enum: [one_to_many, one_to_one, many_to_many]
                    type: string
...
info:
    description:
        "A Java client JAR is available for use with the group ID 'com.liferay', artifact ID
        'com.liferay.object.admin.rest.client', and version '1.0.0'."
    license:
        name: "Apache 2.0"
        url: "http://www.apache.org/licenses/LICENSE-2.0.html"
    title: "Object"
    version: v1.0
openapi: 3.0.1
paths:
...
    "/object-definitions/{objectDefinitionId}/object-relationships":
        get:
            operationId: getObjectDefinitionObjectRelationshipsPage
            parameters:
                - in: path
                  name: objectDefinitionId
                  required: true
                  schema:
                      format: int64
                      type: integer
                - in: query
                  name: page
                  schema:
                      type: integer
                - in: query
                  name: pageSize
                  schema:
                      type: integer
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: "#/components/schemas/ObjectRelationship"
                                type: array
                        application/xml:
                            schema:
                                items:
                                    $ref: "#/components/schemas/ObjectRelationship"
                                type: array
            tags: ["ObjectRelationship"]
        post:
            operationId: postObjectDefinitionObjectRelationship
            parameters:
                - in: path
                  name: objectDefinitionId
                  required: true
                  schema:
                      format: int64
                      type: integer
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/ObjectRelationship"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/ObjectRelationship"
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ObjectRelationship"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/ObjectRelationship"
            tags: ["ObjectRelationship"]
...
    "/object-relationships/{objectRelationshipId}":
        delete:
            operationId: deleteObjectRelationship
            parameters:
                - in: path
                  name: objectRelationshipId
                  required: true
                  schema:
                      format: int64
                      type: integer
            responses:
                204:
                    content:
                        application/json: {}
                        application/xml: {}
            tags: ["ObjectRelationship"]